Split page node utils further
