+++ src/utils/nodes/page/page.ts
@@ -9,7 +9,6 @@ import { EditorState } from "@tiptap/pm/state";
 import { PAGE_NODE_NAME } from "../../../constants/page";
 import { NodePosArray } from "../../../types/node";
 import { Nullable } from "../../../types/record";
-import { getParentNodePosOfType } from "../node";
 import { isPosMatchingEndOfPageCondition, isPosMatchingStartOfPageCondition } from "../../pagination";
 
 /**
@@ -61,72 +60,6 @@ export const collectPageNodes = (doc: PMNode): NodePosArray => {
     });
 
     return pageNodes;
-};
-
-/**
- * Get the page node (parent of the current node) position.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {number} The position of the page node.
- */
-export const getThisPageNodePosition = (doc: PMNode, pos: ResolvedPos | number): number => {
-    return getParentNodePosOfType(doc, pos, PAGE_NODE_NAME).pos;
-};
-
-/**
- * Get the page node position and the page node itself.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {pagePos: number, pageNode: Node} The position and the node of the page.
- */
-export const getPageNodeAndPosition = (doc: PMNode, pos: ResolvedPos | number): { pagePos: number; pageNode: Nullable<PMNode> } => {
-    if (typeof pos === "number") {
-        return getPageNodeAndPosition(doc, doc.resolve(pos));
-    }
-
-    const pagePos = getThisPageNodePosition(doc, pos);
-    const pageNode = doc.nodeAt(pagePos);
-    if (!isPageNode(pageNode)) {
-        console.warn("No page node found");
-        return { pagePos: -1, pageNode };
-    }
-
-    return { pagePos, pageNode };
-};
-
-/**
- * Get the start of the page position.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {number} The start position of the page.
- */
-export const getStartOfPagePosition = (doc: PMNode, pos: ResolvedPos | number): number => {
-    if (typeof pos === "number") {
-        return getStartOfPagePosition(doc, doc.resolve(pos));
-    }
-
-    const { pagePos } = getPageNodeAndPosition(doc, pos);
-
-    return pagePos;
-};
-
-/**
- * Get the end of the page position.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {number} The end position of the page.
- */
-export const getEndOfPagePosition = (doc: PMNode, pos: ResolvedPos | number): number => {
-    if (typeof pos === "number") {
-        return getEndOfPagePosition(doc, doc.resolve(pos));
-    }
-
-    const { pagePos, pageNode } = getPageNodeAndPosition(doc, pos);
-    if (!pageNode) {
-        return pagePos;
-    }
-
-    return pagePos + pageNode.content.size;
 };
 
 /**
